core/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
psycopg2-binary
alembic
pydantic[email]
passlib
python-jose
requests
jinja2
python-multipart
bcrypt